requirements.txt
pybtex==0.24.0
arxiv==1.4.7
scholarly==1.7.11
semanticscholar==0.4.0
pyzotero==1.5.5
schedule==1.2.0
requests==2.31.0
semantic-kernel==0.4.1.dev0
python-dotenv==1.0.0
python-dotenv